+++ src/fdw.c
@@ -97,6 +97,8 @@ enum FdwModifyPrivateIndex
 	FdwModifyPrivateInsertSQL,
 	/* Integer list of target attribute numbers for INSERT/UPDATE */
 	FdwModifyPrivateTargetAttnums,
+	/* List of target attribute OIDs for INSERT/UPDATE */
+	FdwModifyPrivateTargetAttOids,
 	/* Deparsed name of the result table */
 	FdwModifyPrivateTableName,
 };
@@ -255,6 +257,7 @@ static CHFdwModifyState * create_foreign_modify(EState * estate,
 												Plan * subplan,
 												char *query,
 												List * target_attrs,
+												List * target_oids,
 												char *table_name);
 static void finish_foreign_modify(CHFdwModifyState * fmstate);
 static void prepare_query_params(PlanState * node,
@@ -1204,6 +1207,7 @@ clickhousePlanForeignModify(PlannerInfo * root,
 	Relation	rel;
 	StringInfoData sql;
 	List	   *targetAttrs = NIL;
+	List	   *targetTypes = NIL;
 
 	initStringInfo(&sql);
 
@@ -1222,7 +1226,10 @@ clickhousePlanForeignModify(PlannerInfo * root,
 			Form_pg_attribute attr = TupleDescAttr(tupdesc, attnum - 1);
 
 			if (!attr->attisdropped)
+			{
 				targetAttrs = lappend_int(targetAttrs, attnum);
+				targetTypes = lappend_oid(targetTypes, attr->atttypid);
+			}
 		}
 	}
 
@@ -1258,12 +1265,11 @@ clickhousePlanForeignModify(PlannerInfo * root,
 
 	table_close_compat(rel, NoLock);
 
-
 	/*
 	 * Build the fdw_private list that will be available to the executor.
 	 * Items in the list must match enum FdwModifyPrivateIndex, above.
 	 */
-	return list_make3(makeString(sql.data), targetAttrs, makeString(table_name));
+	return list_make4(makeString(sql.data), targetAttrs, targetTypes, makeString(table_name));
 }
 
 /*
@@ -1280,6 +1286,7 @@ clickhouseBeginForeignModify(ModifyTableState * mtstate,
 	CHFdwModifyState *fmstate;
 	char	   *query;
 	List	   *target_attrs = NULL;
+	List	   *target_oids = NULL;
 	RangeTblEntry *rte;
 	char	   *table_name;
 
@@ -1293,6 +1300,7 @@ clickhouseBeginForeignModify(ModifyTableState * mtstate,
 	/* Deconstruct fdw_private data. */
 	query = strVal(list_nth(fdw_private, FdwModifyPrivateInsertSQL));
 	target_attrs = (List *) list_nth(fdw_private, FdwModifyPrivateTargetAttnums);
+	target_oids = (List *) list_nth(fdw_private, FdwModifyPrivateTargetAttOids);
 	table_name = strVal(list_nth(fdw_private, FdwModifyPrivateTableName));
 
 	/* Find RTE. */
@@ -1311,6 +1319,7 @@ clickhouseBeginForeignModify(ModifyTableState * mtstate,
 #endif
 									query,
 									target_attrs,
+									target_oids,
 									table_name);
 
 	resultRelInfo->ri_FdwState = fmstate;
@@ -1343,6 +1352,7 @@ clickhouseBeginForeignInsert(ModifyTableState * mtstate,
 	TupleDesc	tupdesc = RelationGetDescr(rel);
 	int			attnum;
 	List	   *targetAttrs = NIL;
+	List	   *targetTypes = NIL;
 	StringInfoData sql;
 	char	   *table_name;
 
@@ -1352,7 +1362,10 @@ clickhouseBeginForeignInsert(ModifyTableState * mtstate,
 		Form_pg_attribute attr = TupleDescAttr(tupdesc, attnum - 1);
 
 		if (!attr->attisdropped)
+		{
 			targetAttrs = lappend_int(targetAttrs, attnum);
+			targetTypes = lappend_oid(targetTypes, attr->atttypid);
+		}
 	}
 
 	/*
@@ -1383,6 +1396,7 @@ clickhouseBeginForeignInsert(ModifyTableState * mtstate,
 									NULL,
 									sql.data,
 									targetAttrs,
+									targetTypes,
 									table_name);
 
 	resultRelInfo->ri_FdwState = fmstate;
@@ -1537,6 +1551,7 @@ create_foreign_modify(EState * estate,
 					  Plan * subplan,
 					  char *query,
 					  List * target_attrs,
+					  List * target_oids,
 					  char *table_name)
 {
 	CHFdwModifyState *fmstate;
@@ -1570,7 +1585,7 @@ create_foreign_modify(EState * estate,
 
 	old_mcxt = MemoryContextSwitchTo(PortalContext);
 	fmstate->state = fmstate->conn.methods->prepare_insert(fmstate->conn.conn,
-														   rri, target_attrs, &q, table_name);
+														   rri, target_attrs, target_oids, &q, table_name);
 	MemoryContextSwitchTo(old_mcxt);
 
 	/* Create context for per-query temp workspace. */